Add files via upload
Here I add the four classes used to create and aply the merge sort and insertion sort algorithms. 

The class MergeUnsort has an algorithm who give the user and array sort in descendant order [n, n-1, n-2, n-3, ..., 3, 2, 1], this with the objective of take the times in the worst case using the MergeSort algorithm, you can notice that there is not diference between that two classes more than a > sign.